Fixed entry point and core plugins.
I've always wanted logging capabilities when developing nvim config.
With this setup I do, I'm using lazy to pull in packages for an initial
"core" install. Then managing the loading of structlog manually. This
enables me to have a logging plugin available that can be used without
relying on a pluginmanager. The pluginmanager is just there to ensure
that the plugin is up to date.